+++ docs/CHANGE_LOG.md
@@ -8,6 +8,9 @@
 #### **Binding Options - Custom Triggers:**
 - Added a new binding option called "onDayToolTipRender", which states an event that should be called for custom rendering the tooltip for a specific day.
 
+#### **Public Functions:**
+- Added new public function "renderAll()", which will find all new DOM elements with the "data-heat-options" attribute and render them.
+
 #### **CSS:**
 - Added new ":root" variable "--heat-js-day-color-1-text-color".
 - Added new ":root" variable "--heat-js-day-color-2-text-color".